+++ Backend/ChatService/include/chatservice/TokenService.h
@@ -10,7 +10,7 @@ extern const char* kIssuer;
 extern const std::string kKeysDir;
 extern const std::string kPublicKeyFile;
 
-std::optional<int> verifyTokenAndGetUserId(const std::string& token);
+std::optional<long long> verifyTokenAndGetUserId(const std::string& token);
 
 }  // namespace JwtUtils
